+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Embed style sheets in single-file mode using data URIs rather than `<style>`.
 ### Fixed
 * Handle `srcset` entries without a width or pixel density descriptor.
+* Embed subresources recursively when calling `WebResource.to_data_uri()` on an archive's main resource.
 
 ## [0.3.0] - 2021-07-18
 ### Added